Refactor: 결제 페이지 통합
Refs: #7
