+++ src/App.js
@@ -3,7 +3,7 @@ import GlobalStyle from './styles/global';
 import LoginPage from './pages/loginPage/LoginPage';
 import JoinPage from './pages/joinPage/JoinPage';
 import CartPage from './pages/cartPage/CartPage';
-import PaymentPage from './pages/PaymentPage';
+import PaymentPage from './pages/paymentPage/PaymentPage';
 import NotFound from './pages/notFoundPage/NotFoundPage';
 import ProductDetailPage from './pages/ProductDetailPage';
 import SellerCenterPage from './pages/SellerCenterPage';
